Moved ConfigMPS to Config in general
