+++ tests/test_mps.py
@@ -10,7 +10,7 @@ from pytket.circuit import Circuit, Qubit, OpType  # type: ignore
 from pytket.pauli import Pauli, QubitPauliString  # type: ignore
 from pytket.extensions.cutensornet.states import (
     CuTensorNetHandle,
-    ConfigMPS,
+    Config,
     MPS,
     MPSxGate,
     MPSxMPO,
